backend/requirements.txt
Flask==2.3.3
Flask-CORS==3.0.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
PyJWT==2.7.0
